gcp failover attempt
First crack at writing a failover protocol for when the connection to the gcp database is lost.

If the update_db function can't update firestore, then it'll instead save everything to JSON formatted file locally.

Added in logic for checking for a JSON file with the same testID at test start-up. If one exists, check to see whether that's newer, or the GCP data is newer. Use whichever one is newer to load test parameters